graphql settings, an error left

+++ src/main/java/com/example/OrdforandeLista/service/TagService.java
@@ -6,6 +6,8 @@ import com.example.OrdforandeLista.repositories.TagRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+
 @Transactional
 @Service
 public class TagService {
@@ -24,9 +26,12 @@ public class TagService {
     }
 
     public void deleteTag(Long id) {
+        if (!tagRepository.existsById(id)) {
+            throw new IllegalArgumentException("Tag with ID '" + id + "' does not exist");
+        }
         tagRepository.deleteById(id);
-        ;
     }
+
 
     public Tag updateTag(Long id, Tag tag) {
         return tagRepository.findById(id).map(dbTag -> {
@@ -35,4 +40,9 @@ public class TagService {
         }).orElseThrow(()-> new IllegalArgumentException("Kunde inte hitta taggen"));
 
     }
+
+    public List<Tag> getTagsByIds(List<Long> ids) {
+        return tagRepository.findAllById(ids);
+    }
+
 }
